+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Tolerate missing `name` attribute when sorting module descriptors. Refs UID-200.
 * Remove display of mod-configuration values. Refs UID-182.
 * Include mod-settings permissions for access to tenant-wide values. Refs UID-206.
+* Display mod-scheduler's timers. Refs UID-208.
 
 ## [10.0.0](https://github.com/folio-org/ui-developer/tree/v10.0.0) (2025-03-17)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v9.0.0...v10.0.0)